dodawanie ocen z poziomu modala z dynamicznym pobieraniem uczniów

+++ MemoriesBack/MemoriesBack/MemoriesBack/DTO/GradeDetailDTO.cs
@@ -3,7 +3,7 @@
     public class GradeDetailDTO
     {
         public int Id { get; set; }
-        public int Grade { get; set; }
+        public double Grade { get; set; }
         public string Type { get; set; }
         public string IssueDate { get; set; }
         public string Description { get; set; }
@@ -11,7 +11,7 @@
         public string TeacherName { get; set; }
         public string ClassName { get; set; }
 
-        public GradeDetailDTO(int id, int grade, string type, string issueDate, string description, string studentName, string teacherName, string className)
+        public GradeDetailDTO(int id, double grade, string type, string issueDate, string description, string studentName, string teacherName, string className)
         {
             Id = id;
             Grade = grade;
