+++ MemoriesBack/MemoriesBack/MemoriesBack/DTO/GradeRequest.cs
@@ -5,7 +5,9 @@
         public int StudentId { get; set; }
         public int TeacherId { get; set; }
         public int ClassId { get; set; }
-        public int Grade { get; set; }
+
+        public double Grade { get; set; } 
+
         public string? Description { get; set; }
         public string? Type { get; set; }
